+++ sway/commands/layout.c
@@ -154,8 +154,8 @@ static struct cmd_results *cmd_layout_auto(swayc_t *container, int argc, char **
 				}
 				if (inc) {
 					for (int i = container->nb_master;
-					i >= 0 && i < container->children->length
-						&& i != (int)container->nb_master + inc;) {
+							i >= 0 && i < (int)container->children->length
+								&& i != (int)container->nb_master + inc;) {
 						((swayc_t *)container->children->items[i])->height = -1;
 						((swayc_t *)container->children->items[i])->width = -1;
 						i += inc > 0 ? 1 : -1;
