+++ sway/security.c
@@ -48,7 +48,7 @@ struct feature_policy *alloc_feature_policy(const char *program) {
 	if (!validate_ipc_target(program)) {
 		return NULL;
 	}
-	for (int i = 0; i < config->feature_policies->length; ++i) {
+	for (size_t i = 0; i < config->feature_policies->length; ++i) {
 		struct feature_policy *policy = config->feature_policies->items[i];
 		if (strcmp(policy->program, "*") == 0) {
 			default_policy = policy->features;
@@ -76,7 +76,7 @@ struct ipc_policy *alloc_ipc_policy(const char *program) {
 	if (!validate_ipc_target(program)) {
 		return NULL;
 	}
-	for (int i = 0; i < config->ipc_policies->length; ++i) {
+	for (size_t i = 0; i < config->ipc_policies->length; ++i) {
 		struct ipc_policy *policy = config->ipc_policies->items[i];
 		if (strcmp(policy->program, "*") == 0) {
 			default_policy = policy->features;
@@ -143,7 +143,7 @@ static const char *get_pid_exe(pid_t pid) {
 struct feature_policy *get_feature_policy(const char *name) {
 	struct feature_policy *policy = NULL;
 
-	for (int i = 0; i < config->feature_policies->length; ++i) {
+	for (size_t i = 0; i < config->feature_policies->length; ++i) {
 		struct feature_policy *p = config->feature_policies->items[i];
 		if (strcmp(p->program, name) == 0) {
 			policy = p;
@@ -164,7 +164,7 @@ uint32_t get_feature_policy_mask(pid_t pid) {
 	uint32_t default_policy = 0;
 	const char *link = get_pid_exe(pid);
 
-	for (int i = 0; i < config->feature_policies->length; ++i) {
+	for (size_t i = 0; i < config->feature_policies->length; ++i) {
 		struct feature_policy *policy = config->feature_policies->items[i];
 		if (strcmp(policy->program, "*") == 0) {
 			default_policy = policy->features;
@@ -181,7 +181,7 @@ uint32_t get_ipc_policy_mask(pid_t pid) {
 	uint32_t default_policy = 0;
 	const char *link = get_pid_exe(pid);
 
-	for (int i = 0; i < config->ipc_policies->length; ++i) {
+	for (size_t i = 0; i < config->ipc_policies->length; ++i) {
 		struct ipc_policy *policy = config->ipc_policies->items[i];
 		if (strcmp(policy->program, "*") == 0) {
 			default_policy = policy->features;
@@ -197,7 +197,7 @@ uint32_t get_ipc_policy_mask(pid_t pid) {
 uint32_t get_command_policy_mask(const char *cmd) {
 	uint32_t default_policy = 0;
 
-	for (int i = 0; i < config->command_policies->length; ++i) {
+	for (size_t i = 0; i < config->command_policies->length; ++i) {
 		struct command_policy *policy = config->command_policies->items[i];
 		if (strcmp(policy->command, "*") == 0) {
 			default_policy = policy->context;
